Break away hll specific functionality into seperate bot file
I plan to store *only* the discord logic in commands.py for slash
commands.

Other logic can be parsed elsewhere, including the RCON client.
